Adds tools and increases download timeout
Adds `jq` and `packer` to the `tools` directory for use in the build script.
Also, includes the `tools` directory in the `.gitignore` file.
The build script now checks for the existence of these tools and downloads them if they are missing.
Additionally, increases the `http_timeout` for ISO downloads from 2h to 4h
to avoid timeout issues during image creation.